Code/requirements.txt
ultralytics
opencv-python
torch
torchvision
tqdm
Pillow